requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
flask-wtf>=1.2.0
authlib>=1.3.0
werkzeug>=3.0.0
requests>=2.31.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
gunicorn>=22.0.0
psycopg2-binary>=2.9.9
